rag_support_bot/requirements.txt
langchain
langchain-community
langchain-openai
langgraph
chromadb
pypdf
python-dotenv
tiktoken
sentence-transformers
